DERBY-3304
When a SQL exception is thrown, make sure that rollback caused by it closes all the resultsets 
irrespective of whether they return rows or not. This cleanup was not happening for
CallableStatementResultSet. To fix this, in CallableStatementResultSet class, I have changed 
the no-op cleanup() method to call close(). Without this, the locks held by the resultsets 
created inside the Java procedure method were not getting released.

I have added a test case to make sure that this code change is tested. I have created a
Java procedure which creates a dynamic resultset, a local resultset and then does an 
insert which will cause duplicate key exception. As part of rollback for exception, Derby
closes the dynamic resultset and local resultset along with the CallableStatementResultset.
And the test case is able to drop the tables which were used by the dynamic and local
resultset without running into locking issues.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@631108 13f79535-47bb-0310-9956-ffa450edef68