DERBY-3094: Grouping by expressions causes NullPointerException
This change modifies GroupByNode.addUnAggColumns to process the expressions
in the GROUP BY list in descending order of complexity, as measured by the
number of column references in the GROUP BY list element. This ensures
that when we are matching up expressions in the SELECT list with expressions
in the GROUP BY list, that we choose the best match, and don't erroneously
match a sub-expression.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@632221 13f79535-47bb-0310-9956-ffa450edef68