DERBY-2109 Security Principals and Permission classes to support system level permissions in Derby through standard Java security policy.
Committed a sub-set of patch 12 related to the security code without enabling any functionality.
Note that the SystemPrivilegesPermissionTest is not added into the _Suite (as the complete patch does)
as I think it will fail under J2ME. It passes with classes but fails with jars but that is to be expected
as the Derby's permissions classes are not yet added into the jars, this will occur once they are used.

Patch contributed by: Martin Zaun Email: martin dot zaun at sun dot com

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@632419 13f79535-47bb-0310-9956-ffa450edef68