DERBY-3330 committed on behalf of Anurag Shekhar
Committed modified versions of patch derby-3330v13.diff and 
db2Compatibility-v2.diff.  The modifications are mostly changed and/or
added comments along with some code formatting changes to make the 
new code match the surrounding code style.

This checkin adds the functionality to
create unique constraints on a key that contains one or more nullable
columns.  This constraint will all inserts of all keys that contain
one or more columns with the null value.  All keys that contain no columns
with null values are constrained to be unique in the table.  

The underlying implementation for this new type of unique constraint uses
the existing btree non-unique index with a small modification to do 
checking at insert time to provide the described unique behavior for null
and non-null keys.  The sorter has also been modified to provide this
support for creating the index.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@633560 13f79535-47bb-0310-9956-ffa450edef68