DERBY-3310 ASSERT in MergeSort.checkColumnTypes() disallow legal type conversions
Changed copyTypesAndLengthsToSource to not change the type of the
underlying expression,  because that same expression is pointed to
by the child result set's result column list.  Only NormalizeResultSetNode
should have a   type that is different than its expresssion.
See Army's write up attached to DERBY-3310 for a detailed analysis of the
 role of copyTypesAndLengthsToSource.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@636872 13f79535-47bb-0310-9956-ffa450edef68