DERBY-3373: SQL distinct and order by needed together.
Derby was issuing the error message "The ORDER BY clause may not specify
an expression, since the query specifies DISTINCT" in situations where
the message was not appropriate. An example of a query that should have
been accepted, but was being rejected, is:

   select name from person order by lower(name)

This patch backs out part of the changes that went in with DERBY-2351;
specifically, this patch reverts the "if" test which was rejecting any
DISTINCT query which had a pulled-up ORDER BY expression. Not all such
expressions are invalid, as the above example shows. Since the "if"
statement cannot distinguish between valid expressions and invalid ones,
it is better to revert to the pre-DERBY-2351 behavior.

A few new tests are also added to demonstrate the behavior.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@637529 13f79535-47bb-0310-9956-ffa450edef68