DERBY-3489: Error message XRE04 does not include the right port number. Contributed by V Narayanan
* java/engine/org/apache/derby/impl/store/replication/net/ReplicationMessageTransmit.java
The constructor has been modified to accept the SlaveAddress object instead of a host name
and port number as was happening previously.

* java/engine/org/apache/derby/impl/store/replication/net/ReplicationMessageReceive.java
Modified the constructor to accept a SlaveAddress object instead of a host name and port number.
Removed the getHostName() and getPort() functions, these functions seemed superfluous. They are
no longer used in the SlaveController, they are used in only once place in the receiver when an
exception was being thrown.

* java/engine/org/apache/derby/impl/store/replication/slave/SlaveController.java
slavehost and slaveport are no longer used (SlaveAddress object is instead used).
introduced two functions getHostName and getPortNumber here that return the hostName
and portNumber from SlaveAddress. 

* java/engine/org/apache/derby/impl/store/replication/master/MasterController.java
slavehost and slaveport are no longer used (SlaveAddress object is instead used).
introduced two functions getHostName and getPortNumber here that return the hostName
and portNumber from SlaveAddress.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@644742 13f79535-47bb-0310-9956-ffa450edef68