migrating revision 645665 from trunk into 10.4 codeline (DERBY-3320).
The commit comments were as follows
Fix for the test that was added for DERBY-3320. The test was using DriverManager.getConnection
but DriverManager is not available in JSR 169 environment. This caused the test to fail
when run in JSR 169 environment. I have fixed the test to use testUtil.getConnection
which is smart enough to use DriverManager or DataSource depending on what VM the test
is getting run in.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.4@645706 13f79535-47bb-0310-9956-ffa450edef68