DERBY-3347: ERROR XSDB3: Container information cannot change once written
(merged from trunk, revision 647091)

On JVMs that support the NIO API, multiple threads may perform I/O
operations concurrently on the same data file. As long as these
operations go through the page cache, only a single thread performs
I/O on a single page at any given time. The data files can also be
accessed by the container cache, which accesses space that it borrows
on the first page in the file. Since these accesses don't go through
the page cache, a mechanism is needed to prevent concurrent access
that page.

This patch makes reading and writing of the first page in a file
synchronize on the container object. Since access to the borrowed
space on the first page also is synchronized on the container,
concurrent I/O on the first page is prevented. (On JVMs that don't
support the NIO API, all page accesses synchronize on the container.)


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.4@647310 13f79535-47bb-0310-9956-ffa450edef68