DERBY-3603: 'IN' clause ignores valid results.
Patch contributed by A B (qozinx at gmail dot com)

Some queries using multi-valued IN clauses were not returning the right
results. An example of a query which was processed incorrectly is:

    select count(*) FROM spike.accounts account, spike.admin_units admin_unit,
        spike.bookings booking
    WHERE booking.child_id = 2 AND
        admin_unit.admin_unit_id IN (1,21) AND
        booking.booking_date_time_out >= 20080331000000 AND
        booking.booking_date_time_in <= 20080406235900 AND
        account.account_id = booking.account_id AND
        admin_unit.admin_unit_id = account.admin_unit_id; 

The issue involves the behavior of MultiProbeTableScanResultSet when it
goes to re-open the scan; under certain circumstances, it was failing to
reset the probing state, and so was performing the probing incorrectly,
using only partial portions of the IN list values. For example, in the
above query, there were certain rows which were only tested against the
value "admin_unit_id = 21"; the "admin_unit_id = 1" case was skipped.

MultiProbeTableScanResultSet.reopenCore() was using a heuristic test to
distinguish between the two cases of:

   * A - The first is for join processing. In this case we have
   * a(nother) row from some outer table and we want to reopen this
   * scan to look for rows matching the new outer row.
   *
   * B - The second is for multi-probing. Here we want to reopen
   * the scan on this table to look for rows matching the next value
   * in the probe list.

The patch modifies this code so that the caller passes in a boolean flag
to specify which case is occurring, which avoids the problem thinking that
it was in case "B" when in fact it was actually in case "A".
  


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@648492 13f79535-47bb-0310-9956-ffa450edef68