DERBY-3613: SELECT DISTINCT with GROUP BY produces wrong results
Certain combinations of DISTINCT and GROUP BY in the same query were
producing incorrect results. Duplicate rows were appearing in the
results because the query was including all of the GROUP BY columns
in the evaluation of the DISTINCT clause, not just the columns that
were explicitly specified to be DISTINCT.

For example, in the query:

  select distinct a, b from t group by a, b, c

Derby was including two separate rows in the result which had the same
value for columns a and b, but had different values for column c.

Internally, GroupByList.bindGroupByColumns() was generating the
extra column(s) from the group by list into the select's result
column list, but this processing should not be performed when the
query specifies distinct, because adding extra columns to the
set of distinct columns changes the results.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@650728 13f79535-47bb-0310-9956-ffa450edef68