DERBY-2457: Use of column aliases in GROUP BY / HAVING clauses may fail
Some use of column aliases in group by / having clauses can cause queries 
to fail with error 42X04. The queries can sometimes be made to work by 
also aliasing the table or rewriting the query to use a subselect.

After analyzing the script, and studying the standard, we came to feel
that Derby was behaving correctly according to the standard. While
column aliases are valid in the ORDER BY clause, they are not valid
in the GROUP BY and HAVING clauses. Instead, Derby currently correctly
enforces the standard's requirement that the underlying column name
be used in these clauses.

This change updates the Derby GroupByTest.java test program to contain
the examples from the reproduction script, demonstrating that Derby's
behavior is correct, and adding to the body of GROUP BY test cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@655947 13f79535-47bb-0310-9956-ffa450edef68