DERBY-2457: Use of column aliases could simplify queries
Derby does not permit the use of column aliases in WHERE clauses. When
a column is given an alias name, that name can only be used in the
ORDER BY clause, not in the WHERE clause (nor in the GROUP BY or
HAVING clauses -- see DERBY-2457).

This change adds several new tests to select.sql, demonstrating the
current Derby behavior, which we believe to be correct according
to the SQL standard.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@656929 13f79535-47bb-0310-9956-ffa450edef68