DERBY-3625
Fixed SYSCS_UTIL.SYSCS_INPLACE_COMPRESS_TABLE() bug where during the 
defragment phase one part of the code would think there was enough room to
move a record, but because the new record id on the destination page would
take more room than on the source page the move actually would not succeed.
In this case an XSDA3 errror would be thrown and the compress would fail.

This would cause intermittent errors in the nightly concateTests.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@671605 13f79535-47bb-0310-9956-ffa450edef68