DERBY-3625
Adding space check to case where we try to move row to a new page.  I could
not come up with a repro for this case, but seems safer to check for space
than not.  The case the code is trying to avoid is a free page that has been
reclaimed with a nextid that requires more space than the row on the source
page.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@672834 13f79535-47bb-0310-9956-ffa450edef68