DERBY-3319: Logical connections do not check if a transaction is active on close
Make sure that connections created by the different kinds of data
sources throw an exception on close if they are active. This is done
to get the same behaviour as with connections returned by
DriverManager.

Don't throw exception for connections with auto-commit on (since
they'll auto-commit the transaction as part of the close, and
therefore won't leave uncommitted operations around) or connections
that are part of an XA transaction (since those transactions can still
be committed/aborted via the associated XAResource after the closing
of the connection).


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@675870 13f79535-47bb-0310-9956-ffa450edef68