DERBY-3764: Union Query fail on Derby 10.4.1.3
Only look at the visible columns when checking that the expressions in
a UNION query are compatible. This prevents
ArrayIndexOutOfBoundsException from being thrown when preparing some
queries that contain GROUP BY clauses. It is only a partial fix, since
some of the queries run into other problems later in the execution.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@676819 13f79535-47bb-0310-9956-ffa450edef68