DERBY-961: Error 22001 can include 'XX-RESOLVE-XX' which should be removed
In some cases, error 22001 can come up as follows:

   A truncation error was encountered trying to shrink
   CHAR () FOR BIT DATA 'XX-RESOLVE-XX' to length 40.

This message makes it appear as if there is a column called XX-RESOLVE-XX,
however, this string is hard-coded in SQLBinary.java, and should be removed.

The intent of the XX-RESOLVE-XX message parameter is to hold the actual
value which could not be shrunk to fit in the column. For a SQLBinary
data type, the actual value may not be human-readable, and so cannot
be displayed.

This patch changes the XX-RESOLVE-XX parameter to the message to
"(Binary data value not displayed)", which should give a better hint to
the person who received this error message about what was going on. 

The patch also adds a new test case to BlobMemTest to force the code to
emit the 22001 message from SQLBinary, to ensure that this code path
is run in our tests.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@678769 13f79535-47bb-0310-9956-ffa450edef68