DERBY-3408: Wrong message when using SHOW in ij
This change enhances the ij tool so that, if it detects that a syntax
error has occurred, it will now emit an additional message suggesting
some useful resources that might help the user. The new message is:

    Issue the 'help' command for general information on IJ command syntax.
    Any unrecognized commands are treated as potential SQL commands and executed directly.
    Consult your DBMS server reference documentation for details of the SQL syntax supported by your server.

The change includes updates to a number of regression test output files
which include intentional syntax errors in ij scripts.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@694954 13f79535-47bb-0310-9956-ffa450edef68