DERBY-3872
Merging 705037 from trunk into 10.3 release where this bug was introduced. Copying the commit comments from 705037 here for reference.

The NPE in this jira entry was caused by the missing overwrite of accept() method in IndexToBaseRowNode. 
Because of the missing code, the additional layer of VirtualColumn node over ResultColumn was not 
happening for the where clause in HAVING. Once the accept method was added to IndexToBaseRowNode, the 
VirtualColumn on top of the ResultColumn got the correct resultset number associated with it and at the 
code generation time, we start referencing the correct resultset rather than the one associated with 
the JOIN clause. Thanks a ton to Army and Bryan on this jira entry for their help. 

I have added a test case for this in lang/GroupByTest.java 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@705660 13f79535-47bb-0310-9956-ffa450edef68