DERBY-3936: Add CharacterStreamDescriptor. Added a descriptor object for streams representing characters. The following information is tracked by the descriptor:  - if the stream can/should be buffered  - if the stream is position aware (and can reposition itself)  - the current byte position  - the current character position (starting at 1, and special value BEFORE_FIRST)  - byte length  - character length  - maximum allowed length (an imposed limit)  - data offset (in case of headers/metadata in the stream)
Note that the descriptor don't handle encoding. Currently all streams are assumed to use the modified UTF-8 encoding (which is Derby's on disk format).

This is the first revision, more pieces of information may be added later.
The intended use of the descriptor is to pass enough information to allow readers/decoders to configure themselves properly at initialization time. The descriptor object is immutable, and it does not keep it's state in synch with the stream.
Patch file: derby-3936-1b-CharacterStreamDescriptor.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@718936 13f79535-47bb-0310-9956-ffa450edef68