DERBY-3981: Improve distribution of hash codes in SQLBinary and SQLChar
Added a performance test which shows the worst-case behaviour of the
current hash function. The test performs SELECT DISTINCT on a CHAR(5)
column and on a CHAR(5) FOR BIT DATA column. The current hash function
maps all the rows in the test to the same bucket and the hash table
used for elimination of duplicates becomes a linked list.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@730689 13f79535-47bb-0310-9956-ffa450edef68