DERBY-4006: ALTER COLUMN .. WITH DEFAULT NULL does not change the default
The code in ModifyColumnNode was having trouble distinguishing between
the case where the user didn't specify a default at all, versus the
case where they explicitly specified a default value of NULL. 

ModifyColumnNode wants to be able to ask if "no default value was specified",
so this change adds a boolean instance variable "keepCurrentDefault" to
the superclass ColumnDefinitionNode, and initializes it to TRUE if no
default clause was specified. Then ModifyColumnNode uses this variable
to decide whether to preserve the current default value when re-building
the column's definition.

Several new test cases are also added.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@733401 13f79535-47bb-0310-9956-ffa450edef68