DERBY-3270 Delayed (on-demand) creation of current user schema makes select from view belonging to other schema fail.
Fixed with patch derby-3270-dhw-3. This temporarily sets the default
current schema back to what it was at the time of definition of the
schema. This need also exsts in other places, so a general mechanism
to do this was factored in and used in all such places.

New testcase added.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@735084 13f79535-47bb-0310-9956-ffa450edef68