DERBY-4013 Allow standard SQL syntax: ALTER TABLE ALTER [COLUMN] <col> SET DEFAULT <default>
Patch derby-4014d, which adds the standard SQL syntax SET DEFAULT and DROP DEFAULT, in
addition to the non-standard syntax we already have.

Note that we are not 100% in compliance, since DROP DEFAULT does not
fail as it should when there is no existing default - it is a noop right now.

A new test, AlterColumnTest, is also added.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@737774 13f79535-47bb-0310-9956-ffa450edef68