DERBY-3420: Remove obsolete -ca option from ij.
The -ca command-line-option for the ij command is obsolete and no longer
needed, since the ij connection URL can now accept all the necessary
connection attribute information.

This change removes the implementation of the -ca option from the ij code,
and removes the option from the ij "usage" message, and from the test output.

There is a release note attached to DERBY-3420, noting that the option
has been removed, and that users of this option (hopefully there are none)
should instead pass connection attributes in the URL.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@737948 13f79535-47bb-0310-9956-ffa450edef68